also protect delete and put method

+++ pages/api/flashcards/[id].js
@@ -1,18 +1,26 @@
 import Flashcard from "@/db/models/flashcard";
 import dbConnect from "@/db/connect";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(request, response) {
+  const session = await getServerSession(request, response, authOptions);
   await dbConnect();
   const { id } = request.query;
 
   if (request.method === "PUT") {
     try {
-      const flashcardData = request.body;
-      const flashcardToUpdate = await Flashcard.findByIdAndUpdate(
-        id,
-        flashcardData,
-        { new: true }
-      );
+      if (session) {
+        const flashcardData = request.body;
+        const flashcardToUpdate = await Flashcard.findByIdAndUpdate(
+          id,
+          flashcardData,
+          { new: true }
+        );
+      }
+      if (!session) {
+        response.status(401).json({ status: "Not authorized" });
+      }
 
       if (!flashcardToUpdate) {
         response.status(404).json({ status: "Flashcard not found" });
@@ -29,7 +37,12 @@ export default async function handler(request, response) {
 
   if (request.method === "DELETE") {
     try {
-      const deleted = await Flashcard.findByIdAndDelete(id);
+      if (session) {
+        const deleted = await Flashcard.findByIdAndDelete(id);
+      }
+      if (!session) {
+        response.status(401).json({ status: "Not authorized" });
+      }
 
       if (!deleted) {
         response.status(404).json({ status: "Flashcard not found" });
